pyproject.toml
[project]
name = "incident-copilot"
version = "0.1.0"
description = "AI-powered incident classification and analysis copilot with RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
